Added ToCommand for requests

+++ PetFamily.Backend/src/PetFamily.Application/Features/Volunteers/HardDelete/HardDeleteVolunteerHandler.cs
@@ -1,38 +1,45 @@
 using CSharpFunctionalExtensions;
 using Microsoft.Extensions.Logging;
+using PetFamily.Application.Extentions;
+using PetFamily.Application.Interfaces;
 using PetFamily.Domain.Shared;
 
 namespace PetFamily.Application.Features.Volunteers.HardDelete;
 
-public class HardDeleteVolunteerHandler
+public class HardDeleteVolunteerHandler : ICommandHandler<Guid, DeleteVolunteerCommand>
 {
     private readonly IVolunteersRepository _volunteersRepository;
+    private readonly DeleteVolunteerCommandValidator _validator;
     private readonly ILogger<HardDeleteVolunteerHandler> _logger;
 
     public HardDeleteVolunteerHandler(
         IVolunteersRepository volunteersRepository,
-        ILogger<HardDeleteVolunteerHandler> logger
-        )
+        DeleteVolunteerCommandValidator validator,
+        ILogger<HardDeleteVolunteerHandler> logger)
     {
         _volunteersRepository = volunteersRepository;
+        _validator = validator;
         _logger = logger;
     }
-    
-    public async Task<Result<Guid, Error>> Handle(
-        DeleteVolunteerRequest request,
+
+    public async Task<Result<Guid, ErrorList>> HandleAsync(
+        DeleteVolunteerCommand command,
         CancellationToken cancellationToken = default)
     {
-        var volunteerResult = await _volunteersRepository.GetById(request.VolunteerId, cancellationToken);
+        var validationResult = await _validator.ValidateAsync(command, cancellationToken);
+        if (!validationResult.IsValid)
+            return validationResult.ToErrorList();
+
+        var volunteerResult = await _volunteersRepository.GetById(command.VolunteerId, cancellationToken);
         if (volunteerResult.IsFailure)
-            return volunteerResult.Error;
-        
+            return volunteerResult.Error.ToErrorList();
+
         await _volunteersRepository.Delete(volunteerResult.Value, cancellationToken);
-        
+
         await _volunteersRepository.Save(volunteerResult.Value, cancellationToken);
-        
+
         _logger.LogInformation("Volunteer with id: {VolunteerId} was HARD deleted", volunteerResult.Value.Id);
 
         return volunteerResult.Value.Id.Value;
     }
-    
 }
